pi todo try ended

+++ pipccold.py
@@ -14,7 +14,12 @@
 """
 
 path="/media/usbhdd/colddownload/"
+#path="./download/"
 roomid="cold"
+streamtype='middle'
+
+setHowLong=True
+pikll=True
 howlong=60*30 #30min
 
 roomapi='http://open.douyucdn.cn/api/RoomApi/room/'
@@ -23,7 +28,7 @@ myemail="zaowuworld@163.com"
 
 # test
 #roomid="kpc"
-#path="./download/"
+
 
 
 
@@ -79,30 +84,35 @@ def testroomstatus(roomid):
         return
 
 def savestream(roomid,streams,objstr):
-    if 'source' in streams.keys():
+    if streamtype in streams.keys():
+        p=streamtype
+    elif 'source' in streams.keys():
         p='source'
-    elif 'best' in streams.keys():
-        p='best'
-    elif 'middle' in streams.keys():
-        p='middle'
     else:
         p=streams.keys()[0]
     logging.info('save '+p+'#'+objstr)
     now=datetime.datetime.now().strftime('_%I_%M')
     filename=objstr+now+'.mp4'
-    cmd='livestreamer -o "'+path+filename+'" '+roomurl+roomid+' '+p
+    cmd='livestreamer -o "'+path+filename+'" '+roomurl+roomid+' '+p#+' &'
     logging.info('do '+cmd)
     shell=subprocess.Popen(cmd,shell=True)
-    time.sleep(howlong)
-    t=threading.Thread(target=main)
-    t.start()
-    time.sleep(10)
-    shell.kill()
-    kll=subprocess.Popen('kill -9 '+str(shell.pid+1),shell=True)
-    logging.info('save end '+str(shell.pid)+' '+filename)
+
+    if setHowLong:
+        # time limit
+        time.sleep(howlong)
+        t=threading.Thread(target=main)
+        t.start()
+        time.sleep(10)
+        shell.kill()
+        logging.info('save end '+str(shell.pid)+' '+filename)
+        if pikll:
+            # pi
+            kll=subprocess.Popen('kill -9 '+str(shell.pid+1),shell=True)
+    
 
 
 def main():
+    logging.info('do main')
     try:
         obj=testroomstatus(roomid)
